Switch to PostGIS instead of plain PostgreSQL
The first step is to edit ``dotcloud.yml``, and replace our
plain PostgreSQL with PostGIS. Technically, PostGIS is just
PostgreSQL with additional geographic extensions. It provides
new SQL types, functions, and also a bunch of tables that
are used by PostGIS to store projection information (allowing
you to add your custom projections without recompiling or
even editing local files), as well as metadata about your
own geographic tables.

..

If you're new to PostGIS, we recommend checking out
http://postgis.refractions.net/ to get a primer about
how it works. It's not strictly necessary to work with
GeoDjango, but it can help -- just like knowing SQL is
not mandatory to work with Django, but will make you
a better Django developer.