Add geographic app
We will now write the (possibly) simplest GeoDjango app ever.
It defines a model with a ``PointField`` and a name. Think of it
as a way to record famous landmarks. Of course, GeoDjango also has
other types, allowing to use lines, polygons, as well as collections
of those basic types. Available fields and their options are
described in the documentation at
https://docs.djangoproject.com/en/dev/ref/contrib/gis/model-api/.

Also, we will allow this app to show up in the Django admin site.
GeoDjango provides a special geographic admin, that displays
geographic fields on a map, and allows you to edit them, just like
you would edit a regular text or numeric field.

..

At that point, your first GeoDjango app is ready: push it to
DotCloud! At the end of the push, you will see the application
URL. Add "/admin" at the end of the URL, login with "admin"
and "password", and try to add something in the "Places" table!